Simplify model initialisation
The API observer now produces a stream of List<Track> that get incrementally bigger as the waveforms are downloaded. Thus we can just keep assigning the current track list to be the result of the call().

This is useful as we can keep recycling the old content until we know that new content is available. Especially useful when the connection fails.

Note - the call() mechanism doesn't provide an obvious way to know if you are first emission on a subscription, and you can't just clear the track list unconditionally.